split emotion class and eye class into two files
